requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.0
requests>=2.31.0
langchain>=0.1.17
openai>=1.30.1
langchain-community>=0.0.34
